+++ apps/bot/package.json
@@ -14,7 +14,6 @@
     "@ssb/orm": "workspace:*",
     "@ssb/ui": "workspace:*",
     "dotenv": "^16.3.1",
-    "express": "^4.18.2",
     "node-schedule": "^2.1.1",
     "ts-node": "^10.9.1"
   },
